server/src/main/roomman/requirements.txt
fastapi
pydantic
uvicorn
motor
asyncio
pika
aio_pika
python-dotenv
fastapi_security
celery
flower
boto3 
python-multipart